Updated routes to use render method.
