+++ client/public/index.html
@@ -12,6 +12,7 @@
 
 <body>
   <div id="app"></div>
-<script type="text/javascript" src="bundle.js"></script></body>
+  <script type="text/javascript" src="bundle.js"></script>
+</body>
 
 </html>
